requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
mysql-connector-python==9.1.0
numpy==2.2.1
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.3.0
sqlalchemy==2.0.36
pymysql==1.1.1